2.AgenticAI/requirements.txt
langchain
langgraph
langchain-core
langchain-community
langchain-openai
ipykernel
typing_extensions
typing
python-dotenv
langchain-groq
arxiv
wikipedia
langgraph-cli[inmem]
